Separating traits by usage.
