+++ src/Traits/CanBlock.php
@@ -4,14 +4,6 @@ namespace Rennokki\Befriended\Traits;
 
 trait CanBlock
 {
-    public function blockers($model = null)
-    {
-        return $this->morphToMany(($model) ?: $this->getMorphClass(), 'blockable', 'blockers', 'blockable_id', 'blocker_id')
-                    ->withPivot('blocker_type')
-                    ->wherePivot('blocker_type', ($model) ?: $this->getMorphClass())
-                    ->wherePivot('blockable_type', $this->getMorphClass());
-    }
-
     public function blocking($model = null)
     {
         return $this->morphToMany(($model) ?: $this->getMorphClass(), 'blocker', 'blockers', 'blocker_id', 'blockable_id')
