+++ src/Traits/CanFollow.php
@@ -4,14 +4,6 @@ namespace Rennokki\Befriended\Traits;
 
 trait CanFollow
 {
-    public function followers($model = null)
-    {
-        return $this->morphToMany(($model) ?: $this->getMorphClass(), 'followable', 'followers', 'followable_id', 'follower_id')
-                    ->withPivot('follower_type')
-                    ->wherePivot('follower_type', ($model) ?: $this->getMorphClass())
-                    ->wherePivot('followable_type', $this->getMorphClass());
-    }
-
     public function following($model = null)
     {
         return $this->morphToMany(($model) ?: $this->getMorphClass(), 'follower', 'followers', 'follower_id', 'followable_id')
